requirements.txt
flask
gunicorn
pandas
psycopg2-binary
flask-bcrypt
plotly
ollama
numpy
